+++ tests/immediate.rs
@@ -1,22 +1,20 @@
 mod common;
 
-use common::{ShareableI2c, ADDR, I2C_BUS};
+use common::{ShareableI2c, ADDR, I2C_MUTEX};
 
-use lazy_static::lazy_static;
+use embedded_hal_bus::i2c::MutexDevice;
+use once_cell::sync::Lazy;
 use std::sync::Mutex;
 
 use pca9535::{Expander, Pca9535Immediate, Register};
 
 pub type ImmediateExpander = Pca9535Immediate<ShareableI2c>;
 
-lazy_static! {
-    static ref EXPANDER: Mutex<ImmediateExpander> = {
-        let i2c_bus = *I2C_BUS.lock().unwrap();
-        let expander = Pca9535Immediate::new(i2c_bus.acquire_i2c(), ADDR);
+static EXPANDER: Lazy<Mutex<ImmediateExpander>> = Lazy::new(|| {
+    let expander = Pca9535Immediate::new(MutexDevice::new(&I2C_MUTEX), ADDR);
 
-        Mutex::new(expander)
-    };
-}
+    Mutex::new(expander)
+});
 
 #[test]
 fn read_write_byte() {
@@ -191,10 +189,11 @@ mod standard {
 
 #[cfg(test)]
 mod pin {
-    use super::common::{Pca9535GPIO, ShareableI2c, ADDR, I2C_BUS, RPI_GPIO};
+    use super::common::{Pca9535GPIO, ShareableI2c, ADDR, I2C_MUTEX, RPI_GPIO};
     use super::ImmediateExpander;
 
-    use lazy_static::lazy_static;
+    use embedded_hal_bus::i2c::MutexDevice;
+    use once_cell::sync::Lazy;
     use serial_test::serial;
     use std::sync::Mutex;
 
@@ -211,32 +210,31 @@ mod pin {
         >,
     >;
 
-    lazy_static! {
-        static ref IO_EXPANDER: IoExpander<ShareableI2c, ImmediateExpander, Mutex<ImmediateExpander>> = {
-            let i2c_bus = *I2C_BUS.lock().unwrap();
-            let expander = Pca9535Immediate::new(i2c_bus.acquire_i2c(), ADDR);
+    static IO_EXPANDER: Lazy<
+        IoExpander<ShareableI2c, ImmediateExpander, Mutex<ImmediateExpander>>,
+    > = Lazy::new(|| {
+        let expander = Pca9535Immediate::new(MutexDevice::new(&I2C_MUTEX), ADDR);
 
-            IoExpander::new(expander)
+        IoExpander::new(expander)
+    });
+    static PCA9535_GPIO: Lazy<Pca9535Gpio> = Lazy::new(|| {
+        let pca9535_gpio = Pca9535GPIO {
+            _in0_3: ExpanderInputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 3).unwrap(),
+            in0_4: ExpanderInputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 4).unwrap(),
+            _out0_7: ExpanderOutputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 7, PinState::High)
+                .unwrap(),
+            out1_5: ExpanderOutputPin::new(&*IO_EXPANDER, GPIOBank::Bank1, 5, PinState::Low)
+                .unwrap(),
         };
-        static ref PCA9535_GPIO: Pca9535Gpio = {
-            let pca9535_gpio = Pca9535GPIO {
-                _in0_3: ExpanderInputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 3).unwrap(),
-                in0_4: ExpanderInputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 4).unwrap(),
-                _out0_7: ExpanderOutputPin::new(&*IO_EXPANDER, GPIOBank::Bank0, 7, PinState::High)
-                    .unwrap(),
-                out1_5: ExpanderOutputPin::new(&*IO_EXPANDER, GPIOBank::Bank1, 5, PinState::Low)
-                    .unwrap(),
-            };
 
-            Mutex::new(pca9535_gpio)
-        };
-    }
+        Mutex::new(pca9535_gpio)
+    });
 
     #[test]
     #[serial(immediate_pin)]
     fn input_pin_is_high() {
         let rpi_gpio = &mut *RPI_GPIO.lock().unwrap();
-        let pca9535_gpio = PCA9535_GPIO.lock().unwrap();
+        let mut pca9535_gpio = PCA9535_GPIO.lock().unwrap();
 
         rpi_gpio.out0_4.set_high();
 
@@ -247,7 +245,7 @@ mod pin {
     #[serial(immediate_pin)]
     fn input_pin_is_low() {
         let rpi_gpio = &mut *RPI_GPIO.lock().unwrap();
-        let pca9535_gpio = PCA9535_GPIO.lock().unwrap();
+        let mut pca9535_gpio = PCA9535_GPIO.lock().unwrap();
 
         rpi_gpio.out0_4.set_low();